Module quill::update

Structs

ZigDownloadInfo

ZigRelease

ZigReleaseDownloadInfo

Enums

QuillVersion

Statics

APP_USER_AGENT

Functions

download_archive_or_exit

If a version is provided, this function assumes that the tarball contains a version.txt file, and that the version should match the given expected version. Archive types tar.gz, tar.xz and zip are supported. If zip is provided, we assume that we’re downloading the zig compiler, and special-case logic is used to remove the top-level folder from the archive.

download_artifact

If unpack_inner_folder is true, the artifact contains exactly one folder with the same name, which will be unpacked.

download_self

download_text_or_exit

process_update